pyproject.toml
[project]
name = "neurde"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "h5py>=3.15.1",
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.8",
    "pytorch-optimizer>=3.9.0",
    "pyyaml>=6.0.3",
    "rich>=14.3.1",
    "scipy>=1.17.0",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu130", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[dependency-groups]
dev = [
    "autoflake>=2.3.1",
    "black>=26.1.0",
    "pytest>=9.0.2",
]

[tool.autoflake]
check = false
